Edit migrations and add product model with controller and tests

+++ database/migrations/2025_11_30_225116_create_products_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->decimal('price', 10, 2);
-            $table->unsignedTinyInteger('total_stock');
-            $table->unsignedTinyInteger('available_stock');
+            $table->unsignedInteger('total_stock');
+            $table->unsignedInteger('available_stock');
             $table->timestamps();
 
             $table->index('available_stock');
